Use the stackTrace field for group key based Compose traces
Changes how the `DiagnosticComposeException` is created for group key traces. Instead of using the message to show the trace, appends each frame as a `StackTraceElement`. This makes the exception instance more correct while also preserving the trace if the message is stripped.

Test: ErrorTraceTests
Relnote: Fill in the stack trace for `DiagnosticComposeException` in release mode.
Change-Id: Ib76cba026417427c9bb7eac21b3b3e3a942ae956