Fix animating out entry going blank
With the fix in aosp/3784586, the new scene now replaces the old scene immediately upon gesture back/onBack complete. Since new scene has the same key as the old scene, new scene entries overrides the old entries in the `scenesToRnderableEntry` map.

For example if popping from old scene [a, b, c] to new scene [a, b], the scenes to rednerable entry map now only contains entries [a, b] as 'perssimible to render' for the top-most scene. Since the popped entry 'c' is no longer included, the container for c's content is now blank as it animates out as the top-most scene.

To resolve this, the scene to render map is refactored into the scene to exclude map - instead of designating which entries a scene can render, it now designates which entries a scene cannot render.

In the example above when scene [a, b, c] pops to [a, b], the outgoing scene will not be able to render a, b since it is now reserved for the incoming scene, but it can still render c, thereby preventing c from becoming blank while animating out.

Bug: n/a
Test: ./gradlew navigation3:navigation3-ui:androidConnectedCheck

Change-Id: I099a1b9ee43e9ff573ee9d938b8f610a04ac66f6