Add `captureRemoteDocument` for generation from a Service
Adds a new public suspend function `captureRemoteDocument` to capture a RemoteCompose document by rendering a Composable in a virtual display.

This utility function can be used for testing or for generating documents on the fly. It takes a context, display info, profile, and the Composable content, and returns the resulting ByteArray.

Test: Included in CL
Change-Id: I7fd2bac01751b53b4955a450cb40eabd1345670e