Fix marquee behavior in RTL
The previous behavior was incorrect for RTL: it equated RTL behavior with
the behavior of LTR, with a reversed direction. This shows itself as a bug
where end of the RTL content is where the marquee animation freezes,
rather than the start of the RTL content.

The new behavior now splits out direction from layout direction, to allow
for all 4 intentional behaviors. I also refactored the logic slightly so
that the LTR and RTL behaviors both make more sense.

Relnote: Fixes marquee behavior for RTL layout directions

Fix: 303194380
Test: Updated BasicMarqueeTest

Change-Id: Ib8be3ad95cba2d7b0199e0c21c371ef0e14f4f46