Make sure that AGP and Lint versions are correctly set in integration branch
Prior to this change only buildSrc/settings.gradle was reading version
overrides which made sure that we use the correct version of AGP and
lint for the main build tasks, but for custom lint check tests we used
libs.lintTesting which did not get overridden because we did not read the overrides
in root/settings.gradle. This change fixes that.

Test: None
Change-Id: I49cc89bc5dde0d6fb9d885311545c161d794aabc