Update ExperimentalDetector for new modeling of property annotations
Previously, annotations applied to a property could be found through `UField.uAnnotations`, even if the property didn't technically apply to the field use-site. This has been "fixed", so `uAnnotations` is more technically correct, but meant annotations that only apply to the property use-site wouldn't be realized in UAST. To address that, there is now `UField.sourceAnnotations` which will also include property annotations, and `JavaEvaluator.getAllAnnotations` from lint will handle getting the right annotations.

The `UElement.isDeclarationAnnotatedWith` used to take a nullable `JavaEvaluator` and only search meta-annotations if the evaluator was non-null. To be able to use `JavaEvaluator.getAllAnnotations` and for clarity, the evaluator is now required and there is a separate `searchMetaAnnotations` parameter to determine whether to search the meta-annotations.

Bug: 449008759
Test: KotlinAnnotationsDetectorTest, RequiresOptInDetectorTest
Change-Id: I830e132286cf9e2f049fde3ac499c8e1a042f360