Hoist Configuration in AndroidComposeView
This refactors how the Configuration is updated for Compose.

Previously, the configuration was strangely set via a observer callback,
that was set as a side-effect of the composition of
ProvideAndroidCompositionLocals.

Now, the configuration is directly hoisted in AndroidComposeView
in a mutable state, and is observed normally.

This also allows optimizing other derived values, by only updating them
if needed - for example this avoids recreating the Density object on
every configuration change even if the density didn't change.

Fixes: 295175897
Test: Existing and updated

Change-Id: I8b8a8632998ced98931cd662028f92293b1afe6c