Remove RoomOpenHelper and pre-KMP abstract functions.
The RoomOpenHelper is no longer used and this change removes the backwards compatibility path that enabled newer runtime to be compatible with pre-KMP generated and compiled code. This is done as part of the removal of SupportSQLite.

The pre-KMP abstract functions used to maintain old-gen code compatibility are also removed and the replacements that couldn't be abstract are converted to be abstract as they should always be overridden in generated implementation.

Bug: 438041176
Test: Existing
Relnote: Remove RoomOpenHelper and pre-KMP codegen abstract functions that are no longer used.
Change-Id: Iddeaf974d2a185c1b57cf6589d11561cf33695c6