Fix placeholder span sizing to use non-linear font scaling
Previously, PlaceholderSpan used the font scale as a direct multiplier to convert from sp to px.
This used to be correct, but with non-linear font scaling, the font scale is no longer a proportional
value and instead there is non-linear behavior.

With this change, the density is passed through to the PlaceholderSpan to be able to properly convert a
value in sp to a value in pixels.

Relnote: PlaceholderSpan now correctly uses non-linear font scaling for sizing

Fixes: 324462728
Test: Updated tests

Change-Id: Id2ead31ea5f608657561eff3748b75abf21b73af