Make ProtoLayoutScope to be NonNull field
This is because clients who create MaterialScope with
`materialScopeWithResources` have the scope but they'd still
need to do !!. Going forward, it is expected that majority of
clients would use `materialScopeWithResources` if they need to
use the ProtoLayoutScope, so this simplifies clients code.

There is also a method to check presence of the scope.

Bug: 450067019
Test: N/A
Relnote: "ProtoLayoutScope is now NonNull field in MaterialScope, with additional hasProtoLayoutScope check."
Change-Id: I1858f9cb4ee014f2147ced6487368b5b891b8b80