Closely track LayoutCoordinates updates and reflect changes in overlay
This CL tracks changes to the local position of the LayoutCoordinates
by accessing them in the PlacementScope. This ensures whenever the
position of the accessed coordinates changes, we get a callback
to update any dependent position. This allows content to
be rendered in overlay to promptly reflect any position update
to the layout and render that change in the overlay.

Fixes: 389037232
Test: New tests added

Change-Id: Ie745896fe2063f9d7c13141885494c9340e413a8