Configure busy_timeout before sync and journal mode
This change tries to alleviate the 'Error code: 5, message: database is locked' that has been reported during the opening of new connections by setting the busy_timeout as soon as a connection is opened. Thus making the opening procedure more resistant to locked issued during the synchronous and journal mode configurations.

Bug: 380088809
Bug: 450598290
Test: Existing
Change-Id: Ia76d3c7400f268461dfcf4f2df209e109370c2d8