[2Page] Add SinglePageLayoutStrategy
Implements the SinglePageLayoutStrategy for vertically displaying PDF pages.

This new LayoutStrategy arranges pages in a single vertical column with configurable spacing. It calculates the positions of pages and determines the visible page range within the viewport. It also handles horizontal centering for pages smaller than the view and proportional scrolling for pages that are wider.

Bug: 449727004
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I40b21f4e55d5957f8207bb8be03d833e91b95331