Fix: Enable ignored Row/Column layout tests
Enables `testRow_withMinMainAxisSize` and `testColumn_withMinMainAxisSize`.

These tests verify that a Row or Column with a weighted child can expand to fill a minimum size constraint. They were previously ignored due to a measurement issue where a composable with a `weight` modifier inside a wrapping `Row` or `Column` could not be correctly measured.

Using `IntrinsicSize.Max` allows the `Row`/`Column` to correctly measure its children in this scenario, fixing the tests.

Fixes: 438913990

Test: Re-enabled ignored tests
Change-Id: If05aa76df576985c252adbef5d4250c425974a27