Split Retain.kt into separate source files
Reorganizes Retain.kt to no longer be a "god file". This organization
made more sense when Retain was part of the runtime module, but
splitting it apart

This is strictly an organization change. There are no behavior changes,
or source differences. Moving the factory functions here is a
source-compatible, binary-incompatible change.

Test: N/A
Relnote: N/A
Change-Id: Id665306681eb7ab13a934154eda31a8f2bcc3359