Fail build if host test reports are missing
This change updates the `AndroidXHostTestFailureHandlerPlugin` to verify that a test report exists for each failed host test.

Previously, if a host test hung, it would not generate a report. The build would treat this as a simple test failure and becasue of the --continue flag, not fail the build.

Now, the plugin checks for a corresponding report zip in the `dist/host-test-reports` directory. If the report is missing, the build fails with an error indicating the test likely hung.

BUG: 451564625
TESTED: Ran  ./gradlew :compose:runtime:runtime:allHostTests --continue and it failed as expected as :compose:runtime:runtime:wasmJsBrowserTest.zip does notg. I tested on the change being prior to aosp/3797570, which surfaced this issue

Change-Id: I3c84c136b750dd4ed37bd4b4f20002d646e01d2c