Change RequestUnbufferedDispatchTest to use StandardTestDispatcher
Changes RequestUnbufferedDispatchTest to explicitly use a StandardTestDispatcher instead
of the currently default unconfined one.

This change initially caused the buffered test to fail, as all 1000 motion events were
handled in a single frame.

With the updated approach, I use a Choreographer.FrameCallback to keep track of "real"
frames as encountered by the MotionEvent dispatching system.

Bug: 254115946
Test: Updated

Change-Id: Ie909b98199ab32202abe0ce840c6803880d052ab