Use optimized hash and equals in resources
Previously, we implemented hashCode and equals, but for Dynamic data
protos we just used toByteArray, which is not efficient, but we don't
expect too much usage of it.

However, now with aosp/3691292 there is a proper way of hashing
dynamic data by looking and its content all the way to the last
leaf, and we should use that in Resources and Triggers too.

Additionally, fix equals and hashCode methods in DynamicProtoHasEquals
- the problem is that initial switch case per type was on the Dynamic
builder object but it was comparing instance to proto classes and was
never true. Fix is to have the first pass and switch on dynamic builder
and then internal methods would translate it to proto and to the
comparison of inner cases.

Bug: 431678176
Test: Existing ones
Change-Id: Ie128f66caeb4e8969bfc70225790509d850b5658