Restore entity visibility on non-resize events
If `mAutoHideContent` is enabled, an entity's content is hidden (alpha set to 0) at the start of a resize operation. Previously, if a non-resize event (like a move) occurred before the resize completed, the content would remain hidden.

This change ensures that if a non-resize `ReformEvent` is received while the content is hidden for resizing, the entity's alpha is restored to its original value.

Bug: b/444448478

Test: Added unit test, tested manually with integrated scenecore test
app.

Change-Id: I4a0d6667699cfb3441f051684e8ec2a8a7cc398f