change SwitchButton's tick icon from square edges to rounded edges
Previously, StrokeCap.Butt was used as it didn't have an extension,
so to use .Round for rounded edges, the length/position of the lines
have had to be slightly modified.
Additionally, the animateTick function (and it's dependencies) have been
copied to a new file from materialcore into material3, since this change
only needs to affect material3's SwitchButton (and later Checkbox).
Test: Manual
Bug: 378642590
Relnote: "SwitchButton's Tick icon is now rounded instead of straight"

Change-Id: I31fbfdb376d27e2c2caf019eef39ae6d2673d816