Partially revert window size API changes to workaround test breakages
There are a bunch of clients relying on
DeviceConfigurationOverride.ForecedSize to test against different screen
sizes, which is not correct as ForcedSize is meant to override "local"
size, not window size. However the respective WindowSize overriding API
is not available yet. In order not to break those clients' tests for
now, this CL partially rollback aosp/3792686 to revert the logic to use
WindowInfo.containerDpSize directly, so those tests can still work until
we have a way to migrate them when the WindowSize override API is
available.

Test: existing tests
Change-Id: Ib721f8aaaaf72191d0eb35492ebbaabe0c77d92d