Enforce non-empty use case list for creating SessionConfig
Currently, SessionConfig documents that `useCases` property should not
be empty, but does not enforce it.

This CL adds the logic to enforce it. However, some internal
SessionConfig extensions may not want this behavior and may require the
freedom to pass an empty list of use cases (e.g. LegacySessionConfig).
So, the enforcement is guarded by a `requireNonEmptyUseCases` property
that can be overriden by extension classes within CameraX. The property
is restricted and cannot be overriden by public users without
suppression.

Bug: 452501364
Test: SessionConfigTest
Change-Id: I9ff500bab7febca23802242b65520bd1659da18d