Implement SwipeDismissableSceneStrategy and SwipeToDismissScene
Add a SceneStrategy that will delegate to either:
1. A scene that delegates gesture back handling to SwipeToDismissBox for API <= 35
2. A scene that delegates gesture back handling to NavigationBackHandler (the new PredictiveBack) for API 36+

This CL also adds the SwipeToDismissScene which renders content in a SwipeToDismissBox. Of note, the Scene.key is Unit - this key forms a part of NavDisplay's AnimatedContent.contentKey, which means it prevents AnimatedContent from running transitions when content changes. This ensures that new content will trigger recompositions, but transitions will be left to SwipeToDismissBox.

Bug: 427463127
Test: ./gradlew :wear:compose:compose-navigation3:connectedAndroidTest
Relnote: "Added a SceneStrategy for Wear integration with Navigation3. It can be used with a NavDisplay to display content within a Wear Material compoenent."

Change-Id: I7aa08c636c20977053d99511a1c750293ae13442