Remove requirement for setting useAndroidX
With AGP 9.0-alpha04 it is no longer required to call useAndroidX
property as it is set by default. Instead of checking if it is true, we
should only throw if it is manually set to false now.

RelNote: "Navigation SafeArgs plugin will no longer require setting the
useAndroidX property when being used with AGP 9.0.0-alpha04+."
Test: tested in sample app
Bug: 444746731
Bug: 443106400

Change-Id: I6c3a4eecab1c0927d72742952acdf4aa135b7c7e