Improve DataStore warm read latency by avoiding thread switch in DataStoreImpl.readState unless necessary
Based on the micro-benchmark running on a Pixel 7 test device (clock locked), the minimum, median and p90 latency of DataStore are reduced from HEAD's 53.447, 54.663 and 55.875 microseconds to 38.401, 39.356 and 39.837 microseconds; GuavaDataStore's warm read minimum, median and p90 latency are reduced from HEAD's 96.829, 115.719, 130.167 microseconds to 83.228, 102.397, 105.985 microseconds. This change brings a 28% reduction to the cached data.first() read median latency, and 11.5% reduction to GuavaDataStore's cached read latency.

Bug: 432466011
Test: ./gradlew :datastore:datastore-core:testAndroid
Change-Id: I7817700ea5b20069beb98d8f1d7c2770900d65fc