[2Page] Enable LayoutStrategy for SinglePage layout in PdfViewer.
This change enables the use of `LayoutStrategy` for the `SinglePage` layout in the `PdfViewer` component. This provides a more flexible foundation for page layout that is a prerequisite for implementing a two-page view.

The new strategy is controlled by a feature flag, which has been exposed in the test application to facilitate testing.

This change is a preparatory step toward the larger goal of supporting dual-page layouts in the PDF viewer.

Bug: 452936052
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I4443b0f3d066e790103c692945f2bcae0be6e3f5