Remove auto-added kotlin-stdlib in multi-modules-testapp
The `org.jetbrains.kotlin.android` plugin automatically adds `kotlin-stdlib`
to the `api` configuration, which causes dexer to remove `kotlin/jvm/functions/Function0`

This change adds a workaround to the `multi-modules-testapp` to
manually remove this dependency. This can be removed once the project
migrates to the built-in Kotlin plugin.

BUG: 452908579
TESTED: Ran `executeAppFunction_schemaCreateNoteSerialization_success` and it passes now
Change-Id: Ie82ff7abf6a028e76bfce082c9cc19c36cd41fa6