Rename isFeatureGroupSupported to isSessionConfigSupported
Rename the API name because it should be able to support any types of SessionConfig.

- The original isFeatureGroupSupported accepts a SessionConfig parameter. Actually, it could be used to check the support of a HighSpeedVideoSessionConfig.
- Update the code logic in CameraInfoInternal to check whether the camera can support the extension mode for ExtensionSessionConfig.

The isFeatureGroupSupported was updated as non-experimental API in aosp/3806263. The API is still in 1.6.0-alpha stage and is allowed to be renamed.

Relnote: "Renamed isFeatureGroupSupported to isSessionConfigSupported, and make it support any types of SessionConfig."

Bug: 448781299
Test: CameraInfoDeviceTest && ExtensionSessionConfigTest && FeatureCombinationDeviceTest && ./gradlew :camera:camera-core:checkApi
Change-Id: I2161610a0e067756efd906284bfe4c69a3029c1a