Add SectionedItemTem#AlphabeticalIndexingStrategy
This will replace #allowAlphabeticalIndexing as it has more flexibility.
As the template is not yet released (but already used by a 3p app),
this change only deprecates the old field. The boolean field will be removed before the API is released to prod.

Test: ./gradlew :car:app:app:test --tests "androidx.car.app.model.SectionedItemTemplateTest"
Bug: 410092683
Fix: 454115761
Relnote: "Add SectionedItemTemplate#alphabeticalIndexingStrategy. This is
    a replacement to #isAlphabeticalIndexingAllowed and gives developers
    more options for alphabetical indexing."
Change-Id: Ia164dd88866ec0c8caf99bb5f61b8e0721851a1c