Pin focused AndroidView
Adds automatic pinning support for a focused AndroidView.

Modifier.focusable has out-of-the-box pinning support when a node is focused.

However, AndroidView doesn't use Modifier.focusable,
and instead uses lower level building blocks to aid in the focus management
interop between views and Compose.

Previously, these lower level building blocks did not integrate with the
pinning system, which meant that a focused AndroidView would not be pinned
automatically.

This then led to other side-effects, like reentrant measurement and layout
calls when a focused embedded AndroidView was removed as a result of its node
being reused.

Fixes: 435544101

Change-Id: I4241690a965b0d335e75db27f84e2324c77a9f3e