Migrate auto-close feature to driver APIs.
This change migrates Room's auto-close database functionality (i.e. the setAutoCloseTimeout API from the database builder) to support a Room database that has a driver installed. The SupportSQLite version is then removed. This change does not make the API KMP, but enables driver users to continue using the functionality.

With this change the SupportSQLite API isOpen() is removed from RoomDatabase as it relied on auto-close functionality. A RoomDatabase is considered open / active until an explicit call to close(), so there is no use-case to continue supporting the SupportSQLite API isOpen().

Bug: 316944816
Bug: 438041176
Test: AutoCloserTest, AutoClosingDatabaseTest and others.
Relnote: Remove RoomDatabase.isOpen() a SupportSQLite API.
Change-Id: I77ee4f957557f03ba6bbbcb34b7729ee402d4965