Consolidate pre-packaged options into a config type
This change hides the pre-package configurations from the DatabaseConfiguration and consolidate them into an internal sealed interface. The config is hidden because the functionality and APIs being part of Room 3.0 instead of a more decoupled SQLiteDriver utility is still up to debate. Therefore, hide the configuration for now but continue supporting the actual functionality as it was migrated to driver APIs. Note that the PrepackagedDatabaseCallback is still a public API pending the move of the functionality of out Room.

Bug: 339934813
Bug: 438041176
Test: Existing
Relnote: Remove the pre-package configurations from DatabaseConfiguration.
Change-Id: Icaf0cc99d90e227615ed16d1076ead5282dc6e13