Remove SupportSQLite from Room
This change removes the compatibility mode in Room that enabled the underlying database to be a SupportSQLiteDatabase from a SupportSQLiteOpenHelper. A SQLiteDriver is now required for a RoomDatabase, if none is provided the AndroidSQLiteDriver will be used in Android, for other platforms a database builder will continue to require a setDriver() call.

This change also removes the getOpenHelper() API that was the entry-point to the underlying SupportSQLiteDatabase that enabled database operations outside of a RoomDatabase.

To continue using SupportSQLite APIs with Room obtain a compatibility wrapper from the extension function RoomDatabase.getSupportWrapper().

Bug: 316944352
Bug: 408062492
Bug: 438041176
Test: Existing
Relnote: "Remove Room's SupportSQLite APIs that enabled a RoomDatabase to be backed by a SupportSQLiteDatabase. A SQLiteDriver is now required. The builder API to provide a SupportSQLiteOpenHelper.Factory is removed and the getOpenHelper() in RoomDatabase is also removed. To continue using SupportSQLite APIs with Room get a wrapper instance using RoomDatabase.getSupportWrapper()."
Change-Id: Ie9d17cde0f84bcad1cc9bdee418786eca7678819