Migrate RxJava3 invalidation helper functions to use Flow
The two public utility functions createFlowable() and createObservable() were based on InvalidationTracker.Observable and had not been converted to use the first party Flow APIs from the invalidation tracker. This change migrates them to eliminate usages of the Observer APIs as it is meant to be removed.

Lastly there was a lingering code gen utility function that had not been marked as deprecated, createSingle(), it is no longer used by gen code and so it is removed.

The RxRoom test is deleted but two tests in RxJava3Test are added to cover the utility functions, the tests used Mockito and so the instrumented test is more truthful to the expected behaviour.

Test: Existing
Bug: 438041176
Relnote: "Remove the Room code gen utility function not longer used by the RxJava3 support."
Change-Id: I37422bf9155b1bf64c13ab89ed28fea56e4b0727