Add one more bit for last child offset
And store values divided by LongsPerItem. With those changes, the new max supported lastChildOffset is 1024 instead of previous 170. Also fixed the bug, where if the real offset is larger than max we were iterating the whole LongArray even if the real stored number of items is smaller (happens after item removals and defragmentation).
I was able to get extra space be reducing one bit from id's fields. It seems to be safe to do so as we still can fit ids up to 33,000,000. If in future we realise we need more here we would need to add extra long per item, but I've decided it is not needed yet, so we don't use more memory. Doing so frees one more bit, which I am going to use in upcoming cls for hasCallback flag as another optimization.

Test: updated test to check the logic on larger items count
Change-Id: Ic8beda21a5690e2cd6b6eb670e44dfa96d9b0814