CacheWindow refill update #2
When re-filling the cache window we were using old data, so we sometimes avoided re-caching items that have changed. When the dataset has changed and we need to re-fill the window, we need to clean-up the cached information (prefetched and item sizes) so the algorithm for prefetching will run for all new items. We're also now using the fact that there was a change in the visible item's keys to indicate the need to refill the window.

Test: Updated existing test.
Fixes: 454439658
Relnote: Introduce isCacheWindowRefillFixEnabled to control the roll out of the cache window refill fix.

Change-Id: I4e52baefd57b178834444a60cf43a016ec193e52