CacheWindow refill update #2
When re-filling the cache window we were using old data, so we sometimes avoided re-caching items that have changed. When the dataset has changed and we need to re-fill the window, we need to clean-up the cached information (prefetched and item sizes) so the algorithm for prefetching will run for all new items. We're also now using the fact that there was a change in the visible item's keys to indicate the need to refill the window.

Test: Updated existing test.
Fixes: 454439658
Change-Id: I716ec2adbb230d6292743034e1ea1a2d9d36a766