Fix ThreePaneScaffold crashes when margins are set
When pane margins are set and dragging the pane to reduce its size to be
less than the margin value, the measure policy will crash due to the
incorrect negative pane width calculated by the fact that the margin
"pushes" the left/right edge of the pane over the other edge.

Fixes this by just never using negative width/height values - when the
above situation happens, we should just use zero widths/heights, i.e.,
hiding the relevant pane and let the other pane to account for filling
the space.

Relnote: Fixes ThreePaneScaffold crashes when margins are set
Test: manual
Fixes: 455614909
Change-Id: I141815cdf8330133d43cb61db913b77f43600447