[2Page] Add TwoPageLayoutStrategy for two-page PDF viewing
Implements the TwoPageLayoutStrategy for displaying PDF pages in a dual-page, side-by-side format.

This new LayoutStrategy arranges pages in a two-column layout, which is particularly useful for tablet and desktop form factors. It includes logic to handle the placement of pages, manages the gutter space between them, and determines the visible page range within the viewport. The strategy also accounts for screen orientation and document direction (LTR/RTL) to correctly position the pages.

Bug: 449726934
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I65687962a2acd8e162c4d865bf951cece772b66e