Cycle through session configurations in test app
This change introduces the ability to cycle through a predefined list
of `Config` objects in the `ProjectedTestappActivity`. Center button
is now used to switch to the next configuration in the
list.

This makes it easier to test various ARCore session configurations,
such as enabling/disabling Geospatial mode and device tracking,
without needing to recompile the application for each scenario.

The UI has been updated to display the name of the currently active
configuration and any exceptions that occur during reconfiguration.

The available configurations are:
*   "Geospatial On, 3DoF On": Prints latitude, longitude, altitude, and device pose.
*   "Geospatial Off, 3DoF On": Prints only the device pose.
*   "Geospatial Off, 3DoF Off": Displays the message "All tracking is disabled."
*   "Geospatial On, 3DoF Off": Throws and prints an exception.


Change-Id: Iadb7b9390d0bfd98039d3395f4363c0c30a14edb
Test:  ./gradlew xr:arcore:integration-tests:projected-testapp:installDebug && ./gradlew xr:arcore:arcore-projected:ktCheck && rm local.properties
Bug: b/452091636
RelNote: Geospatial and DeviceTracking are supported on projected